+++ fluxburst/client.py
@@ -5,9 +5,7 @@
 
 import collections
 
-import flux
-import flux.job
-
+import fluxburst.handles as handles
 import fluxburst.selectors as selectors
 import fluxburst.sorting as sorting
 from fluxburst.logger import logger
@@ -20,7 +18,7 @@ class FluxBurst:
     Flux Burst Client
     """
 
-    def __init__(self, handle=None):
+    def __init__(self, handle=None, mock=False):
         """
         Create a new burst client.
 
@@ -32,7 +30,7 @@ class FluxBurst:
         """
         self.reset_selector()
         self.reset_plugins()
-        self.handle = handle or flux.Flux()
+        self.flux = handles.FluxMock(handle) if mock else handles.FluxHandle(handle)
         self.set_ordering(sorting.in_order)
 
     @property
@@ -175,9 +173,7 @@ class FluxBurst:
 
         # Update the KVS (is this possible)?
         # This doesn't currently work, so not doing anything :)
-        kvs = flux.job.job_kvs(self.handle, job["id"])
-        kvs["jobspec"] = job["spec"]
-        kvs.commit()
+        self.flux.update_jobspec(job)
 
     def select_jobs(self):
         """
@@ -188,11 +184,11 @@ class FluxBurst:
         for bursting. See the README / documentation for example info.
         """
         # Keep track of selected burstable jobs by id
-        listing = flux.job.job_list(self.handle).get()
+        listing = self.flux.list_jobs()
         selected = {}
 
         for job in listing.get("jobs", []):
-            info = self.get_job_info(job["id"])
+            info = self.flux.get_job_info(job["id"])
             if not self._job_selector(info):
                 continue
             print(f"🧋️  Job {job['id']} is marked for bursting.")
@@ -207,26 +203,3 @@ class FluxBurst:
 
     def __str__(self):
         return "[flux-burst-client]"
-
-    def get_job_info(self, jobid):
-        """
-        Get job info based on an id
-
-        Also retrieve the full job info and jobspec.
-        This is not yet currently perfectly json serializable, need to
-        handle EmptyObject if that is desired.
-        """
-        fluxjob = flux.job.JobID(jobid)
-        payload = {"id": fluxjob, "attrs": ["all"]}
-        rpc = flux.job.list.JobListIdRPC(self.handle, "job-list.list-id", payload)
-        job = rpc.get_job()
-
-        # Job info, timing, priority, etc.
-        job["info"] = rpc.get_jobinfo().__dict__
-        job["info"]["_exception"] = job["info"]["_exception"].__dict__
-        job["info"]["_annotations"] = job["info"]["_annotations"].__dict__
-
-        # the KVS will have annotations!
-        kvs = flux.job.job_kvs(self.handle, jobid)
-        job["spec"] = kvs.get("jobspec")
-        return job